backend/requirements.txt
Django==3.2.15
django-colorfield==0.7.2
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf-extra-fields==3.2.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
PyJWT==2.1.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
django-extra-fields==3.0.2
 